Fix override if config file already exists from mounting and support APP_BASE_URL env var at runtime

+++ .docker/entrypoint.sh
@@ -2,12 +2,17 @@
 
 set -e
 
-echo "setting config"
+if [ ! -f /usr/share/nginx/html/config.json ]; then
+    echo "Setting config"
+    /app/set-env-vars.js
+    cp /app/config.json /usr/share/nginx/html/config.json
+fi
 
-/app/set-env-vars.js
+if [[ ! -z "$APP_BASE_URL" ]]; then
+    echo "Setting custom APP_BASE_URL"
+    sed -i.bak "s/https:\/\/lite-explorer.aleth.io/$APP_BASE_URL/" /usr/share/nginx/html/index.html
+fi
 
-cp /app/config.json /usr/share/nginx/html/config.json
-
-echo "lite explorer started"
+echo "Lite Explorer started"
 
 exec "$@"
